Release 2.0.0 rc9 * Update paths and models to v2 to show backcompat * Fix endpoints management * Fix tests and update dependencies. * Documentation * Deployment changes

+++ src/Microsoft.Azure.IIoT.OpcUa.Testing/src/Servers/DataAccess/UnderlyingSystemBlock.cs
@@ -127,7 +127,7 @@ namespace DataAccess {
             switch (tag.DataType) {
                 case UnderlyingSystemDataType.Integer1: { tag.Value = (sbyte)0; break; }
                 case UnderlyingSystemDataType.Integer2: { tag.Value = (short)0; break; }
-                case UnderlyingSystemDataType.Integer4: { tag.Value = (int)0; break; }
+                case UnderlyingSystemDataType.Integer4: { tag.Value = 0; break; }
                 case UnderlyingSystemDataType.Real4: { tag.Value = (float)0; break; }
                 case UnderlyingSystemDataType.String: { tag.Value = string.Empty; break; }
             }
@@ -386,7 +386,7 @@ namespace DataAccess {
                             tag.Value = generator.GetRandomInt32();
                         }
                         else {
-                            tag.Value = (int)value;
+                            tag.Value = value;
                         }
 
                         break;
